+++ src/main/java/com/pintoss/auth/core/voucher/application/flow/writer/VoucherIssuerAdder.java
@@ -1,5 +1,6 @@
-package com.pintoss.auth.core.voucher.application.service;
+package com.pintoss.auth.core.voucher.application.flow.writer;
 
+import com.pintoss.auth.core.voucher.application.repository.VoucherIssuerRepository;
 import com.pintoss.auth.core.voucher.domain.VoucherIssuer;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
