+++ src/main/java/com/pintoss/auth/event/SpringVoucherEventPublisher.java
@@ -1,7 +1,7 @@
 package com.pintoss.auth.event;
 
 import com.pintoss.auth.common.event.VoucherPurchaseCompletedEvent;
-import com.pintoss.auth.core.voucher.application.service.VoucherEventPublisher;
+import com.pintoss.auth.core.voucher.application.flow.external.VoucherEventPublisher;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
