+++ src/main/java/com/pintoss/auth/storage/voucher/VoucherRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.pintoss.auth.storage.voucher;
 
 import com.pintoss.auth.core.voucher.domain.Voucher;
-import com.pintoss.auth.core.voucher.application.service.VoucherRepository;
+import com.pintoss.auth.core.voucher.application.repository.VoucherRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
